refactor: service 로직 module-presentation -> module-domain 으로 이동 (#107)
