+++ module-domain/src/main/java/com/depromeet/image/service/ImageGetService.java
@@ -1,0 +1,21 @@
+package com.depromeet.image.service;
+
+import com.depromeet.image.domain.Image;
+import com.depromeet.image.port.in.ImageGetUseCase;
+import com.depromeet.image.port.out.persistence.ImagePersistencePort;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ImageGetService implements ImageGetUseCase {
+    private final ImagePersistencePort imagePersistencePort;
+
+    @Override
+    public List<Image> findImagesByMemoryId(Long memoryId) {
+        return imagePersistencePort.findAllByMemoryIdAndHasUploaded(memoryId);
+    }
+}
