+++ module-domain/src/main/java/com/depromeet/image/service/ImageUpdateService.java
@@ -1,48 +1,40 @@
 package com.depromeet.image.service;
 
 import com.depromeet.exception.BadRequestException;
-import com.depromeet.image.Image;
-import com.depromeet.image.dto.response.ImageUploadResponseDto;
-import com.depromeet.image.repository.ImageRepository;
-import com.depromeet.memory.ImageUploadStatus;
+import com.depromeet.image.domain.Image;
+import com.depromeet.image.domain.ImageUploadStatus;
+import com.depromeet.image.port.in.ImageUpdateUseCase;
+import com.depromeet.image.port.out.command.ImagePresignedUrlCommand;
+import com.depromeet.image.port.out.persistence.ImagePersistencePort;
+import com.depromeet.image.port.out.s3.S3ManagePort;
 import com.depromeet.memory.Memory;
 import com.depromeet.type.image.ImageErrorType;
 import com.depromeet.util.ImageNameUtil;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-import software.amazon.awssdk.services.s3.presigner.S3Presigner;
-import software.amazon.awssdk.services.s3.presigner.model.PresignedPutObjectRequest;
-import software.amazon.awssdk.services.s3.presigner.model.PutObjectPresignRequest;
 
-@Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class ImageUpdateServiceImpl implements ImageUpdateService {
-    private final ImageRepository imageRepository;
-    private final S3Presigner s3Presigner;
-
-    @Value("${spring.cloud.aws.s3.bucket}")
-    private String bucketName;
+public class ImageUpdateService implements ImageUpdateUseCase {
+    private final ImagePersistencePort imagePersistencePort;
+    private final S3ManagePort s3ManagePort;
 
     @Value("${cloud-front.domain}")
     private String domain;
 
     @Override
-    public List<ImageUploadResponseDto> updateImages(Memory memory, List<String> imageNames) {
+    public List<ImagePresignedUrlCommand> updateImages(Memory memory, List<String> imageNames) {
         validateImageIsNotNull(imageNames);
 
-        List<Image> existingImages = imageRepository.findAllByMemoryId(memory.getId());
+        List<Image> existingImages = imagePersistencePort.findAllByMemoryId(memory.getId());
         List<String> existingImageNames = getExistingImageNames(existingImages);
 
-        List<ImageUploadResponseDto> updatedImageNames =
+        List<ImagePresignedUrlCommand> updatedImageNames =
                 getImageUploadResponseDto(memory, imageNames, existingImageNames);
         deleteNonUpdatedImages(existingImages, imageNames);
         return updatedImageNames;
@@ -61,26 +53,26 @@ public class ImageUpdateServiceImpl implements ImageUpdateService {
 
     @Override
     public void changeImageStatus(List<Long> imageIds) {
-        imageRepository.updateByImageIds(imageIds);
+        imagePersistencePort.updateByImageIds(imageIds);
     }
 
     private List<String> getExistingImageNames(List<Image> existingImages) {
         return existingImages.stream().map(Image::getImageName).toList();
     }
 
-    private List<ImageUploadResponseDto> getImageUploadResponseDto(
+    private List<ImagePresignedUrlCommand> getImageUploadResponseDto(
             Memory memory, List<String> imageNames, List<String> existImagesName) {
-        List<ImageUploadResponseDto> imageUploadResponseDtos = new ArrayList<>();
+        List<ImagePresignedUrlCommand> imageUploadResponseDtos = new ArrayList<>();
 
         for (String originImageName : imageNames) {
             String imageName = generateImageName(originImageName);
 
             if (existImagesName.contains(imageName)) continue;
-            String presignedUrl = getPresignedImageUrl(imageName);
+            String presignedUrl = s3ManagePort.getPresignedUrl(imageName);
             Long addedImageId = saveNewImage(originImageName, imageName, memory);
 
-            ImageUploadResponseDto imageUploadResponseDto =
-                    ImageUploadResponseDto.builder()
+            ImagePresignedUrlCommand imageUploadResponseDto =
+                    ImagePresignedUrlCommand.builder()
                             .imageId(addedImageId)
                             .imageName(originImageName)
                             .presignedUrl(presignedUrl)
@@ -107,23 +99,7 @@ public class ImageUpdateServiceImpl implements ImageUpdateService {
                         .memory(memory)
                         .imageUploadStatus(ImageUploadStatus.PENDING)
                         .build();
-        return imageRepository.save(newImage);
-    }
-
-    private String getPresignedImageUrl(String imageName) {
-        PutObjectRequest putObjectRequest =
-                PutObjectRequest.builder().bucket(bucketName).key(imageName).build();
-
-        PutObjectPresignRequest putObjectPresignRequest =
-                PutObjectPresignRequest.builder()
-                        .signatureDuration(Duration.ofMinutes(5))
-                        .putObjectRequest(putObjectRequest)
-                        .build();
-
-        PresignedPutObjectRequest presignedPutObjectRequest =
-                s3Presigner.presignPutObject(putObjectPresignRequest);
-
-        return presignedPutObjectRequest.url().toString();
+        return imagePersistencePort.save(newImage);
     }
 
     private void deleteNonUpdatedImages(
@@ -137,7 +113,7 @@ public class ImageUpdateServiceImpl implements ImageUpdateService {
                         .map(Image::getId)
                         .toList();
         if (!deletedImageIds.isEmpty()) {
-            imageRepository.deleteAllByIds(deletedImageIds);
+            imagePersistencePort.deleteAllByIds(deletedImageIds);
         }
     }
 }
