--- module-presentation/src/main/java/com/depromeet/image/service/ImageDeleteService.java
@@ -1,7 +1,0 @@
-package com.depromeet.image.service;
-
-public interface ImageDeleteService {
-    void deleteImage(Long imageId);
-
-    void deleteAllImagesByMemoryId(Long memoryId);
-}
