--- module-presentation/src/main/java/com/depromeet/image/service/ImageGetService.java
@@ -1,8 +1,0 @@
-package com.depromeet.image.service;
-
-import com.depromeet.image.dto.response.MemoryImagesDto;
-import java.util.List;
-
-public interface ImageGetService {
-    List<MemoryImagesDto> findImagesByMemoryId(Long memoryId);
-}
